requirements.txt
tensorflow_gpu==2.5.0
tensorflow-probability==0.12.0
tensorboardX==2.4.1
scipy==1.5.4
seaborn==0.11.2
matplotlib==3.3.4
gym==0.10.5